requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
celery==5.3.6
channels==4.1.0
channels-redis==4.2.0
Django==5.0.4
django-ckeditor-5==0.2.12
django-debug-toolbar==4.3.0
django-mptt==0.16.0
django-recaptcha==4.0.0
django-redis==5.4.0
django-taggit==5.0.1
oauthlib==3.2.2
pillow==10.3.0
psycopg==3.1.18
psycopg-binary==3.1.18
python-dotenv==1.0.1
redis==5.0.3
requests==2.31.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3

